requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
bcrypt
python-jose[cryptography]
aiosqlite